% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_csv.R
\name{merge_csv}
\alias{merge_csv}
\title{merge csv file}
\usage{
merge_csv(directory, output_file)
}
\arguments{
\item{directory}{the directory where the CSV file is located}

\item{output_file}{The name of the merged csv file}
}
\value{
The merged data frame
}
\description{
this function is mainly used to merge all CSV files in a specified directory
}
\examples{
merge_csv("path/to/csv/files", "output/merged.csv")
}
