% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_excel.R
\name{merge_excel}
\alias{merge_excel}
\title{merge excel file}
\usage{
merge_excel(directory, output_file)
}
\arguments{
\item{directory}{the directory where the excel file is located}

\item{output_file}{the name of the merged excel file}
}
\value{
The merged data frame
}
\description{
this function is mainly used to merge all excel files in a specified directory
}
\examples{
merge_excel("path/to/excel/files", "output/merged.xlsx")
}
